% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetal_zscore.R
\name{fetal_zscore}
\alias{fetal_zscore}
\title{Calculate z-scores for fetal cardiac parameters in Doppler Sono}
\usage{
fetal_zscore(actual, mod_value, card_param, moderator, method, out)
}
\arguments{
\item{actual}{the observed value in mm}

\item{mod_value}{the value of the moderator (see Details)}

\item{card_param}{which cardiac parameters is used (see Details)}

\item{moderator}{one of GA, BPD or femur}

\item{method}{see details}

\item{out}{one of "z" for the acutal z-score or "centile" for the percentile}
}
\value{
z-score or percentile
}
\description{
Calculate z-scores for fetal cardiac parameters in Doppler Sono
}
\details{
\itemize{
\item \emph{mod_value}   for GA: must be in weeks as numeric, continuous value. e.g. 23+4 is given as 23,57 \br
\item \emph{card_param}
in mm, cardiac parametes can be
\itemize{
\item "AoVA" for Aortic valve annulus,
\item "PDA" for patent ductus arteriosus,
\item "Asc Ao", "Dsc Ao" for ascending or descending aorta,
\item "LPA", "RPA", "MPA" for Left \right \main pulmonary artery,
\item "LVL", "RVL" for left or right inlet or length
\item "MVA" for Mitral valve annulus,
\item "PVA" for Pulmonary valve annulus ,
\item "TVA" for Tricuspid valve annulus
\item only Krishnan:
\itemize{
\item "AoI" for Aortic isthmus
}
\item only Schneider:
\itemize{
\item "LVArea", "RVArea", for the circumference of the left or right ventricle, \n
\item "LVEDD", "RVEDD" for end diastolic diameter for left or right ventricle, \n
\item "IVC" for Inferior vena cava,
}
}
\item \emph{moderator}
\itemize{
\item EGA - estimated gestational age in weeks,
\item BPD - biparietal diameter in mm,
\item FL - femur length in mm
\item FW - fetal weight (based on Haddock formula, only available in reference values by Krishan et al. 2016) \n
}
\item \emph{method} one of "Schneider", "Krishnan"
}
}
\examples{
fetal_zscore(actual= 4.8, mod_value=27+3/7, card_param="AoVA", moderator="GA", method="Schneider")
fetal_zscore(actual= 4.8, mod_value=metricGA("27+3"), card_param="AoVA", moderator="GA", method="Schneider")
}
\references{
\enumerate{
\item Krishnan, A. et al. Predictive Models for Normal Fetal Cardiac Structures. Journal of the American Society of Echocardiography 29, 1197–1206 (2016).
\item Schneider, C. et al. Development of Z-scores for fetal cardiac dimensions from echocardiography. Ultrasound in Obstetrics & Gynecology 26, 599–605 (2005).
}
}
