% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricGA.R
\name{metricGA}
\alias{metricGA}
\title{Calculate Gestational Age (GA) in Numeric Format}
\usage{
metricGA(ga_strings)
}
\arguments{
\item{ga_strings}{A character vector of gestational ages in "weeks+days" format
(e.g., c("23+5", "30+2")).}
}
\value{
A numeric vector where each element represents gestational age as "weeks + days/7".
}
\description{
Converts a vector of gestational age strings from "weeks+days" (e.g., "23+5") into numeric format
as "weeks + days/7".
}
\details{
This function parses each input string, splits the weeks and days, and calculates the numeric
value as weeks plus the fraction of days over 7. It ensures that days are between 0 and 6.
}
\examples{
metricGA("23+5") # Returns 23.71
metricGA("40") #Returns 40
metricGA(c("23+5", "30+2", "18+0"))
# Returns: 23.71429, 30.28571, 18.00000

}
